Student Information System
This code creates a simple Student Information System with an input form to add students' data (name, roll number, and grade) and displays the list of students in a table. The data is stored in an array and updated dynamically on the table upon submission. The system is responsive, thanks to the Bootstrap framework used for styling.